beta v2.8.0

+++ uf/modeling/transformer.py
@@ -611,4 +611,4 @@ def positional_encoding(inputs,
         if masking:
             outputs = tf.where(tf.equal(inputs, 0), inputs, outputs)
 
-        return tf.to_float(outputs)
+        return tf.cast(outputs, tf.float32)
